refactor: decompose grep.py into smaller modules
- Created grep_components package with modular structure
- Extracted SearchResult and SearchConfig data classes
- Extracted FileFilter for file filtering functionality
- Extracted PatternMatcher for pattern matching logic
- Extracted ResultFormatter for output formatting
- Reduced main grep.py from 693 to 435 lines
- All modules now under 400 lines (largest is 149 lines)
- All characterization tests still pass
