+++ src/tunacode/tools/grep_components/__init__.py
@@ -1,0 +1,9 @@
+"""
+Grep tool components for modular organization.
+"""
+
+from .file_filter import FileFilter
+from .pattern_matcher import PatternMatcher
+from .search_result import SearchConfig, SearchResult
+
+__all__ = ["PatternMatcher", "FileFilter", "SearchResult", "SearchConfig"]
